pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "i2v"
version = "0.1.0"
description = "SAM-I2V: Promptable Video Segmentation"
authors = [
    {name = "Haiyang Mei", email = "haiyang.mei@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "numpy>=2.0.0",
    "opencv-python>=4.10.0",
    "tqdm>=4.66.0",
    "Pillow>=9.4.0",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    "fvcore>=0.1.5",
    "pandas>=2.2.0",
    "scikit-image>=0.24.0",
    "tensorboard>=2.17.0",
    "pycocotools>=2.0.8",
    "tensordict>=0.5.0",
    "submitit>=1.5.0",
    "natsort",
    "jupyter>=1.0.0",
    "matplotlib>=3.9.0",
    "timm",
    "einops",
]